# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dilshod's API Gateway",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/contract/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Create Contract",
                "parameters": [
                    {
                        "description": "Contract details",
                        "name": "contract",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.CreateContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract created successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ContractResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while creating contract",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contract/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a contract by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Delete Contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ContractResponse"
                        }
                    },
                    "404": {
                        "description": "Contract not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while deleting contract",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contract/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a contract by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get Contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract details",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GetContractResponse"
                        }
                    },
                    "404": {
                        "description": "Contract not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while retrieving contract",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contract/getpdf/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a contract by ID and return it in PDF format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get Contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Contract not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while retrieving contract",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contract/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of contracts with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "List Contracts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by consumer name",
                        "name": "consumer_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by passport seria",
                        "name": "pasport_seria",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of contracts",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GetAllContractResponse"
                        }
                    },
                    "500": {
                        "description": "Error while retrieving contracts",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contract/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Update Contract",
                "parameters": [
                    {
                        "description": "Updated contract details",
                        "name": "contract",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.UpdateContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract updated successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ContractResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while updating contract",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genprotos.ContractResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "genprotos.CreateContractRequest": {
            "type": "object",
            "properties": {
                "consumer_address": {
                    "type": "string"
                },
                "consumer_name": {
                    "type": "string"
                },
                "consumer_passport_serial": {
                    "type": "string"
                },
                "consumer_phone_number": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "passport_image": {
                    "type": "string"
                },
                "status": {
                    "description": "Enum: pending, finished, canceled",
                    "type": "string"
                }
            }
        },
...
        "genprotos.GetAllContractResponse": {
            "type": "object",
            "properties": {
                "all_contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genprotos.GetContractResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.GetContractResponse": {
            "type": "object",
            "properties": {
                "consumer_address": {
                    "type": "string"
                },
                "consumer_name": {
                    "type": "string"
                },
                "consumer_passport_serial": {
                    "type": "string"
                },
                "consumer_phone_number": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "passport_image": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.UpdateContractRequest": {
            "type": "object",
            "properties": {
                "consumer_address": {
                    "type": "string"
                },
                "consumer_name": {
                    "type": "string"
                },
                "consumer_passport_serial": {
                    "type": "string"
                },
                "consumer_phone_number": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "passport_image": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}